factor some expandable layout animation and colors for other fragments to use
